# Issue/1070 Making BlocksOnCylindrical projection tests faster