--- a/src/recon_test/test_blocks_on_cylindrical_projectors.cxx
+++ b/src/recon_test/test_blocks_on_cylindrical_projectors.cxx
@@ -78,23 +78,12 @@ private:
   template <class TProjDataInfo>
   shared_ptr<TProjDataInfo> set_blocks_projdata_info(shared_ptr<Scanner> scanner_sptr);
 
-  template <typename T>
-  void run_symmetry_test(shared_ptr<T> forw_projector1_sptr, shared_ptr<T> forw_projector2_sptr);
-
-  template <typename T>
-  void run_plane_symmetry_test(shared_ptr<T> forw_projector1_sptr, shared_ptr<T> forw_projector2_sptr);
-
+  void run_symmetry_test(ForwardProjectorByBin& forw_projector1, ForwardProjectorByBin& forw_projector2);
+  void run_plane_symmetry_test(ForwardProjectorByBin&forw_projector1, ForwardProjectorByBin& forw_projector2);
   void run_voxelOnCartesianGrid_with_negative_offset();
-
-  template <typename FORW, typename BACK>
-  void run_axial_projection_test(shared_ptr<FORW> forw_projector_sptr, shared_ptr<BACK> bck_projector_sptr);
-
-  template <typename T>
-  void run_map_orientation_test(shared_ptr<T> forw_projector1_sptr, shared_ptr<T> forw_projector2_sptr);
-
-  template<typename Proj1, typename Proj2>
-  void run_projection_test(shared_ptr<Proj1> forw_projector1_sptr, shared_ptr<Proj2> forw_projector2_sptr);
-
+  void run_axial_projection_test(ForwardProjectorByBin& forw_projector, BackProjectorByBin& back_projector);
+  void run_map_orientation_test(ForwardProjectorByBin& forw_projector1, ForwardProjectorByBin& forw_projector2);
+  void run_projection_test(ForwardProjectorByBin& forw_projector1, ForwardProjectorByBin& forw_projector2);
   void run_intersection_with_cylinder_test();
 };
 /*! The following is a function to allow a projdata_info blocksONCylindrical to be created from the scanner.
@@ -127,9 +116,8 @@ BlocksTests::set_blocks_projdata_info(shared_ptr<Scanner> scanner_sptr)
  * The sinogram should have now different values at fixed bin compared the previous image. In this test we are also testing a
  * projdata_info with a negative view offset.
  */
-template <typename T>
 void
-BlocksTests::run_symmetry_test(shared_ptr<T> forw_projector1_sptr, shared_ptr<T> forw_projector2_sptr)
+BlocksTests::run_symmetry_test(ForwardProjectorByBin& forw_projector1, ForwardProjectorByBin& forw_projector2)
 {
 
   //-- ExamInfo
@@ -209,11 +197,11 @@ BlocksTests::run_symmetry_test(shared_ptr<T> forw_projector1_sptr, shared_ptr<T>
   proj_data_info_blocks_sptr = set_blocks_projdata_info<ProjDataInfoBlocksOnCylindricalNoArcCorr>(scannerBlocks_sptr);
   //    now forward-project images
 
-  info(boost::format("Test blocks on Cylindrical: Forward projector used: %1%") % forw_projector1_sptr->parameter_info());
+  info(boost::format("Test blocks on Cylindrical: Forward projector used: %1%") % forw_projector1.parameter_info());
 
-  forw_projector1_sptr->set_up(proj_data_info_blocks_sptr, image1_sptr);
+  forw_projector1.set_up(proj_data_info_blocks_sptr, image1_sptr);
 
-  forw_projector2_sptr->set_up(proj_data_info_blocks_sptr, image2_sptr);
+  forw_projector2.set_up(proj_data_info_blocks_sptr, image2_sptr);
 
   auto projdata1 = std::make_shared<ProjDataInterfile>(exam_info_sptr, proj_data_info_blocks_sptr, "sino1_from_image.hs",
                                                        std::ios::out | std::ios::trunc | std::ios::in);
@@ -221,8 +209,8 @@ BlocksTests::run_symmetry_test(shared_ptr<T> forw_projector1_sptr, shared_ptr<T>
   auto projdata2 = std::make_shared<ProjDataInterfile>(exam_info_sptr, proj_data_info_blocks_sptr, "sino2_from_image.hs",
                                                        std::ios::out | std::ios::trunc | std::ios::in);
 
-  forw_projector1_sptr->forward_project(*projdata1, *image1_sptr);
-  forw_projector2_sptr->forward_project(*projdata2, *image2_sptr);
+  forw_projector1.forward_project(*projdata1, *image1_sptr);
+  forw_projector2.forward_project(*projdata2, *image2_sptr);
   int crystals_in_ring = scannerBlocks_sptr->get_num_detectors_per_ring();
   float bin1_0 = projdata1->get_sinogram(0, 0).at(0 / crystals_in_ring * _PI).at(0);
   float bin1_90 = projdata1->get_sinogram(0, 0).at(90 / crystals_in_ring * _PI).at(0);
@@ -254,9 +242,8 @@ BlocksTests::run_symmetry_test(shared_ptr<T> forw_projector1_sptr, shared_ptr<T>
  *  the forward projected sinogram should show the maximum value at the bin corresponding to the angle phi
  *  equal to the orientation of the plane
  */
-template <typename T>
 void
-BlocksTests::run_plane_symmetry_test(shared_ptr<T> forw_projector1_sptr, shared_ptr<T> forw_projector2_sptr)
+BlocksTests::run_plane_symmetry_test(ForwardProjectorByBin& forw_projector1, ForwardProjectorByBin& forw_projector2)
 {
 
   //-- ExamInfo
@@ -331,21 +318,21 @@ BlocksTests::run_plane_symmetry_test(shared_ptr<T> forw_projector1_sptr, shared_
   shared_ptr<DiscretisedDensity<3, float>> image2_sptr(image2.clone());
   write_to_file("plane30", *image2_sptr);
 
-  info(boost::format("Test blocks on Cylindrical: Forward projector used: %1%") % forw_projector1_sptr->parameter_info());
+  info(boost::format("Test blocks on Cylindrical: Forward projector used: %1%") % forw_projector1.parameter_info());
 
-  forw_projector1_sptr->set_up(proj_data_info_blocks_sptr, image_sptr);
+  forw_projector1.set_up(proj_data_info_blocks_sptr, image_sptr);
 
-  forw_projector2_sptr->set_up(proj_data_info_blocks_sptr, image2_sptr);
+  forw_projector2.set_up(proj_data_info_blocks_sptr, image2_sptr);
 
   auto projdata = std::make_shared<ProjDataInterfile>(exam_info_sptr, proj_data_info_blocks_sptr, "sino1_from_plane.hs",
                                                       std::ios::out | std::ios::trunc | std::ios::in);
 
-  forw_projector1_sptr->forward_project(*projdata, *image_sptr);
+  forw_projector1.forward_project(*projdata, *image_sptr);
 
   auto projdata2 = std::make_shared<ProjDataInterfile>(exam_info_sptr, proj_data_info_blocks_sptr, "sino2_from_plane.hs",
                                                        std::ios::out | std::ios::trunc | std::ios::in);
 
-  forw_projector2_sptr->forward_project(*projdata2, *image2_sptr);
+  forw_projector2.forward_project(*projdata2, *image2_sptr);
 
   int view1_num = 0, view2_num = 0;
   LORInAxialAndNoArcCorrSinogramCoordinates<float> lorB1;
@@ -441,9 +428,8 @@ BlocksTests::run_voxelOnCartesianGrid_with_negative_offset()
  *  The image is forward projected to a sinogram and the sinogram back projected to an image. This image should
  *  be symmetrical along z
  */
-template <typename FORW, typename BACK>
 void
-BlocksTests::run_axial_projection_test(shared_ptr<FORW> forw_projector_sptr, shared_ptr<BACK> bck_projector_sptr)
+BlocksTests::run_axial_projection_test(ForwardProjectorByBin& forw_projector, BackProjectorByBin& back_projector)
 {
 
   //-- ExamInfo
@@ -486,17 +472,17 @@ BlocksTests::run_axial_projection_test(shared_ptr<FORW> forw_projector_sptr, sha
   shared_ptr<DiscretisedDensity<3, float>> bck_proj_image_sptr(image.clone());
   write_to_file("axial_test", *image_sptr);
 
-  info(boost::format("Test blocks on Cylindrical: Forward projector used: %1%") % forw_projector_sptr->parameter_info());
+  info(boost::format("Test blocks on Cylindrical: Forward projector used: %1%") % forw_projector.parameter_info());
 
-  forw_projector_sptr->set_up(proj_data_info_blocks_sptr, image_sptr);
-  bck_projector_sptr->set_up(proj_data_info_blocks_sptr, bck_proj_image_sptr);
+  forw_projector.set_up(proj_data_info_blocks_sptr, image_sptr);
+  back_projector.set_up(proj_data_info_blocks_sptr, bck_proj_image_sptr);
 
   auto projdata = std::make_shared<ProjDataInterfile>(exam_info_sptr, proj_data_info_blocks_sptr, "test_axial.hs",
                                                       std::ios::out | std::ios::trunc | std::ios::in);
 
-  forw_projector_sptr->forward_project(*projdata, *image_sptr);
+  forw_projector.forward_project(*projdata, *image_sptr);
 
-  bck_projector_sptr->back_project(*bck_proj_image_sptr, *projdata, 0, 1);
+  back_projector.back_project(*bck_proj_image_sptr, *projdata, 0, 1);
   write_to_file("back_proj_axial_test", *bck_proj_image_sptr);
 
   int min_z = bck_proj_image_sptr->get_min_index();
@@ -524,10 +510,8 @@ BlocksTests::run_axial_projection_test(shared_ptr<FORW> forw_projector_sptr, sha
  * to look at bin values in the two cases. The bin obtained from the two different projdata will have different coordinates but
  * the same value.
  */
-
-template<typename T>
 void
-BlocksTests::run_map_orientation_test(shared_ptr<T> forw_projector1_sptr, shared_ptr<T> forw_projector2_sptr)
+BlocksTests::run_map_orientation_test(ForwardProjectorByBin& forw_projector1, ForwardProjectorByBin& forw_projector2)
 {
   CPUTimer timer;
   //-- ExamInfo
@@ -624,13 +608,13 @@ BlocksTests::run_map_orientation_test(shared_ptr<T> forw_projector1_sptr, shared
   timer.start();
 
   //    now forward-project images
-  info(boost::format("Test blocks on Cylindrical: Forward projector used: %1%") % forw_projector1_sptr->parameter_info());
+  info(boost::format("Test blocks on Cylindrical: Forward projector used: %1%") % forw_projector1.parameter_info());
   auto projdata1 = std::make_shared<ProjDataInMemory>(exam_info_sptr, proj_data_info_blocks_sptr);
-  forw_projector1_sptr->set_up(proj_data_info_blocks_sptr, image1_sptr);
-  forw_projector1_sptr->forward_project(*projdata1, *image1_sptr);
+  forw_projector1.set_up(proj_data_info_blocks_sptr, image1_sptr);
+  forw_projector1.forward_project(*projdata1, *image1_sptr);
   auto projdata2 = std::make_shared<ProjDataInMemory>(exam_info_sptr, proj_data_info_blocks_reord_sptr);
-  forw_projector2_sptr->set_up(proj_data_info_blocks_reord_sptr, image1_sptr);
-  forw_projector2_sptr->forward_project(*projdata2, *image1_sptr);
+  forw_projector2.set_up(proj_data_info_blocks_reord_sptr, image1_sptr);
+  forw_projector2.forward_project(*projdata2, *image1_sptr);
 
   for (int view = 0; view <= proj_data_info_blocks_reord_sptr->get_max_view_num(); view++)
     {
@@ -665,9 +649,8 @@ BlocksTests::run_map_orientation_test(shared_ptr<T> forw_projector1_sptr, shared
   std::cerr << "-- CPU Time " << timer.value() << '\n';
 }
 
-template<typename Proj1, typename Proj2>
 void
-BlocksTests::run_projection_test(shared_ptr<Proj1> forw_projector1_sptr, shared_ptr<Proj2> forw_projector2_sptr)
+BlocksTests::run_projection_test(ForwardProjectorByBin& forw_projector1, ForwardProjectorByBin& forw_projector2)
 {
   //-- ExamInfo
   auto exam_info_sptr = std::make_shared<ExamInfo>();
@@ -710,23 +693,23 @@ BlocksTests::run_projection_test(shared_ptr<Proj1> forw_projector1_sptr, shared_
   proj_data_info_blocks_sptr = set_blocks_projdata_info<ProjDataInfoBlocksOnCylindricalNoArcCorr>(scannerBlocks_sptr);
 
   //    now forward-project images
-  info(boost::format("Test blocks on Cylindrical: Forward projector used: %1%") % forw_projector1_sptr->parameter_info());
+  info(boost::format("Test blocks on Cylindrical: Forward projector used: %1%") % forw_projector1.parameter_info());
 
-  forw_projector1_sptr->set_up(proj_data_info_blocks_sptr, image1_sptr);
+  forw_projector1.set_up(proj_data_info_blocks_sptr, image1_sptr);
   auto projdata1 = std::make_shared<ProjDataInterfile>(exam_info_sptr, proj_data_info_blocks_sptr, "sino_with_phantom_at_30_0.hs",
                                                        std::ios::out | std::ios::trunc | std::ios::in);
-  forw_projector1_sptr->forward_project(*projdata1, *image1_sptr);
+  forw_projector1.forward_project(*projdata1, *image1_sptr);
   auto projdata2 = std::make_shared<ProjDataInterfile>(exam_info_sptr, proj_data_info_blocks_sptr, "sino_with_phantom_at_25_0.hs",
                                                        std::ios::out | std::ios::trunc | std::ios::in);
-  forw_projector1_sptr->forward_project(*projdata2, *image2_sptr);
+  forw_projector1.forward_project(*projdata2, *image2_sptr);
 
-  forw_projector2_sptr->set_up(proj_data_info_blocks_sptr, image1_sptr);
+  forw_projector2.set_up(proj_data_info_blocks_sptr, image1_sptr);
   auto projdata1_parallelproj = std::make_shared<ProjDataInterfile>(exam_info_sptr, proj_data_info_blocks_sptr, "sino_with_phantom_at_30_0_parallelproj.hs",
                                                  std::ios::out | std::ios::trunc | std::ios::in);
-  forw_projector2_sptr->forward_project(*projdata1_parallelproj, *image1_sptr);
+  forw_projector2.forward_project(*projdata1_parallelproj, *image1_sptr);
   auto projdata2_parallelproj = std::make_shared<ProjDataInterfile>(exam_info_sptr, proj_data_info_blocks_sptr, "sino_with_phantom_at_25_0_parallelproj.hs",
                                             std::ios::out | std::ios::trunc | std::ios::in);
-  forw_projector2_sptr->forward_project(*projdata2_parallelproj, *image2_sptr);
+  forw_projector2.forward_project(*projdata2_parallelproj, *image2_sptr);
 
   //    compare the images: erode all non-zero voxels in ray tracing projections by 1, then sum all voxels in parallelproj projections where the mask is 0
   auto have_same_shape = [](shared_ptr<ProjDataInterfile> projdata, shared_ptr<ProjDataInterfile> projdata_parallel) -> int {
@@ -861,28 +844,28 @@ BlocksTests::run_tests()
   PM->enable_cache(false);
   PM->set_use_actual_detector_boundaries(true);
   PM->set_num_tangential_LORs(5);
-  auto forw_projector1_sptr = std::make_shared<ForwardProjectorByBinUsingProjMatrixByBin>(PM);
-  auto forw_projector2_sptr = std::make_shared<ForwardProjectorByBinUsingProjMatrixByBin>(PM);
-  auto back_projector_sptr = std::make_shared<BackProjectorByBinUsingProjMatrixByBin>(PM);
+  auto forw_projector1 = ForwardProjectorByBinUsingProjMatrixByBin(PM);
+  auto forw_projector2 = ForwardProjectorByBinUsingProjMatrixByBin(PM);
+  auto back_projector = BackProjectorByBinUsingProjMatrixByBin(PM);
 
-  run_symmetry_test(forw_projector1_sptr, forw_projector2_sptr);
-  run_plane_symmetry_test(forw_projector1_sptr, forw_projector2_sptr);
+  run_symmetry_test(forw_projector1, forw_projector2);
+  run_plane_symmetry_test(forw_projector1, forw_projector2);
   run_voxelOnCartesianGrid_with_negative_offset();
-  run_map_orientation_test(forw_projector1_sptr, forw_projector2_sptr);
-  run_axial_projection_test(forw_projector1_sptr, back_projector_sptr);
+  run_map_orientation_test(forw_projector1, forw_projector2);
+  run_axial_projection_test(forw_projector1, back_projector);
   run_intersection_with_cylinder_test();
 
 #ifdef STIR_WITH_Parallelproj_PROJECTOR
   // run the same tests with parallelproj, if available
-  auto forw_projector1_parallelproj_sptr = std::make_shared<ForwardProjectorByBinParallelproj>();
-  auto forw_projector2_parallelproj_sptr = std::make_shared<ForwardProjectorByBinParallelproj>();
-  auto back_projector_parallelproj_sptr = std::make_shared<BackProjectorByBinParallelproj>();
+  auto forw_projector1_parallelproj = ForwardProjectorByBinParallelproj();
+  auto forw_projector2_parallelproj = ForwardProjectorByBinParallelproj();
+  auto back_projector_parallelproj = BackProjectorByBinParallelproj();
 
-  run_symmetry_test(forw_projector1_parallelproj_sptr, forw_projector2_parallelproj_sptr);
-  run_plane_symmetry_test(forw_projector1_parallelproj_sptr, forw_projector2_parallelproj_sptr);
-  run_map_orientation_test(forw_projector1_parallelproj_sptr, forw_projector2_parallelproj_sptr);
-  run_axial_projection_test(forw_projector1_parallelproj_sptr, back_projector_parallelproj_sptr);
-  run_projection_test(forw_projector1_sptr, forw_projector1_parallelproj_sptr);
+  run_symmetry_test(forw_projector1_parallelproj, forw_projector2_parallelproj);
+  run_plane_symmetry_test(forw_projector1_parallelproj, forw_projector2_parallelproj);
+  run_map_orientation_test(forw_projector1_parallelproj, forw_projector2_parallelproj);
+  run_axial_projection_test(forw_projector1_parallelproj, back_projector_parallelproj);
+  run_projection_test(forw_projector1, forw_projector1_parallelproj);
 #endif
 
 }