This PR is about making the BlocksOnCylindrical projection tests faster to avoid AppVeyor timeouts. Current timings with latest commit are below.

-------- Testing Blocks Geometry --------
symmetry test took: 0.947663 seconds
plane symmetry test took: 1.74959 seconds
voxelOnCartesianGrid test took: 0.009056 seconds
axial projection test took: 7.87713 seconds
map orientation test took: 1.40988 seconds
intersection with cylinder test took: 0.122833 seconds

-------- Testing Parallelproj --------
symmetry test took: 1.02611 seconds
plane symmetry test took: 2.07785 seconds
map orientation test took: 0.629038 seconds
axial projection test took: 2.79315 seconds
projection cpu vs. gpu comparison test took: 8.15366 seconds