@markus-jehl 2022-12-29
It does - I actually just piggy-backed on the old ticket to add parallelproj tests and didn't create a new issue (considering this a follow-up job). I updated the title to make it a bit clearer.